Automatically start the user's dbus service.
Check if there is a dbus service definition for the user and start it if
present. This needs to be started before the launcher script exits, so
that the following pam_elogind.so module in the PAM stack can detect the
running dbus at $XDG_RUNTIME_DIR/bus and set DBUS_SESSION_BUS_ADDRESS
into the environment for the remaining session.